requirements.txt
uvicorn==0.32.0
fastapi==0.115.0
uvicorn==0.32.0
python-dotenv
python-jose
passlib
pydantic
